Add files via upload
Simulates mouse movement with predefined coordinates of X and Y, and loops through both of the arrays at the same time. This allows for the mouse to move from point to point. This includes smoothing so the simulated mouse movements do not jump from one coordinate to the next. This smoothing creates a "dragging" effect with the mouse.